# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AEGIS API",
        "description": "Dokumentasi API Project AEGIS",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Login untuk petugas, supervisor, dan warga. Admin tidak bisa login via API ini.",
                "operationId": "391ff20c319bb76c327863f160ad6494",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "petugas@aegis.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "55|ac8pAB1rl3GkCsis..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "Budi Santoso"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "petugas@aegis.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "petugas",
                                                    "enum": [
                                                        "petugas",
                                                        "supervisor",
                                                        "warga"
                                                    ]
                                                },
                                                "tanggal_lahir": {
                                                    "type": "string",
                                                    "example": "1999-05-18"
                                                },
                                                "alamat": {
                                                    "type": "string",
                                                    "example": "Jl. Merdeka No. 1"
                                                },
                                                "no_hp": {
                                                    "type": "string",
                                                    "example": "08123456789"
                                                },
                                                "foto_profil": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/foto_profil.jpg",
                                                    "nullable": true
                                                },
                                                "tanggal_bergabung": {
                                                    "type": "string",
                                                    "example": "2024-01-01"
                                                },
                                                "wilayah_pengawasan": {
                                                    "type": "string",
                                                    "example": "Wilayah A",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Email atau password salah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email atau password salah."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Akun tidak memiliki akses ke aplikasi mobile (role admin)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Akun ini tidak memiliki akses ke aplikasi mobile."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
        "/api/tamu": {
            "get": {
                "tags": [
                    "Tamu"
                ],
                "summary": "Get semua data tamu",
                "operationId": "74a9be7a0f3daa1ef7a8d0559cca92c5",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tamu"
                ],
                "summary": "Tambah data tamu baru",
                "operationId": "106d822703569667befa314045188991",
                "responses": {
                    "201": {
                        "description": "Berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}